Moved GeoCoordinates utilities into their own class
