+++ app/src/main/java/fr/free/nrw/commons/utils/Utils.java
@@ -22,7 +22,6 @@ import java.util.Date;
 import java.util.Locale;
 import java.util.regex.Pattern;
 
-import fr.free.nrw.commons.location.LatLng;
 import timber.log.Timber;
 
 public class Utils {
@@ -77,35 +76,6 @@ public class Utils {
         // Clear previous browser tasks, so that back/exit buttons work as intended.
         customTabsIntent.intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
         customTabsIntent.launchUrl(context, url);
-    }
-
-    /**
-     * Util function to handle geo coordinates. It no longer depends on google maps and any app
-     * capable of handling the map intent can handle it
-     *
-     * @param context The context for launching intent
-     * @param latLng  The latitude and longitude of the location
-     */
-    public static void handleGeoCoordinates(final Context context, final LatLng latLng) {
-        handleGeoCoordinates(context, latLng, 16);
-    }
-
-    /**
-     * Util function to handle geo coordinates with specified zoom level. It no longer depends on
-     * google maps and any app capable of handling the map intent can handle it
-     *
-     * @param context   The context for launching intent
-     * @param latLng    The latitude and longitude of the location
-     * @param zoomLevel The zoom level
-     */
-    public static void handleGeoCoordinates(final Context context, final LatLng latLng,
-        final double zoomLevel) {
-        final Intent mapIntent = new Intent(Intent.ACTION_VIEW, latLng.getGmmIntentUri(zoomLevel));
-        if (mapIntent.resolveActivity(context.getPackageManager()) != null) {
-            context.startActivity(mapIntent);
-        } else {
-            ViewUtil.showShortToast(context, context.getString(R.string.map_application_missing));
-        }
     }
 
     /**
